# more micro optimizations