Hello @bbakerman @andimarek,

I did some more profiling with YourKit and found these new "opportunities":

![image](https://user-images.githubusercontent.com/5798698/102658236-797bc400-4177-11eb-95c8-854471f8b1bb.png)

JMH benchmark:

master
```
Benchmark                                    Mode  Cnt   Score   Error  Units
BenchMark.benchMarkSimpleQueriesThroughput  thrpt   15  47.091 ± 1.469  ops/s
BenchMark.benchMarkSimpleQueriesAvgTime      avgt   15  21.572 ± 1.113  ms/op
```

with FetchedValue fix
```
Benchmark                                    Mode  Cnt   Score   Error  Units
BenchMark.benchMarkSimpleQueriesThroughput  thrpt   15  47.069 ± 0.646  ops/s
BenchMark.benchMarkSimpleQueriesAvgTime      avgt   15  20.776 ± 0.804  ms/op
```

with CacheKey fix
```
Benchmark                                    Mode  Cnt   Score   Error  Units
BenchMark.benchMarkSimpleQueriesThroughput  thrpt   15  52.207 ± 1.457  ops/s
BenchMark.benchMarkSimpleQueriesAvgTime      avgt   15  19.707 ± 0.459  ms/op
```

IMHO is a small but important improvement for all users of the library!