@dfa1 2020-12-22
@bbakerman `PropertyFetcherBenchMark` is also improving on my laptop.

Master branch:
```
# JMH version: 1.25
# VM version: JDK 1.8.0_77, Java HotSpot(TM) 64-Bit Server VM, 25.77-b03
# VM invoker: /Library/Java/JavaVirtualMachines/jdk1.8.0_77.jdk/Contents/Home/jre/bin/java
# VM options: -javaagent:/Users/dfa/Library/Application Support/JetBrains/Toolbox/apps/IDEA-U/ch-0/203.5981.155/IntelliJ IDEA.app/Contents/lib/idea_rt.jar=62641:/Users/dfa/Library/Application Support/JetBrains/Toolbox/apps/IDEA-U/ch-0/203.5981.155/IntelliJ IDEA.app/Contents/bin -Dfile.encoding=UTF-8
# Warmup: 2 iterations, 5 s each, 3 calls per op
# Measurement: 3 iterations, 10 s each, 4 calls per op
# Timeout: 10 min per iteration
# Threads: 1 thread, will synchronize iterations
# Benchmark mode: Throughput, ops/time
# Benchmark: benchmark.PropertyFetcherBenchMark.benchMarkThroughputDirectClassHierarchy

# Run progress: 0.00% complete, ETA 00:06:40
# Fork: 1 of 5
objc[12979]: Class JavaLaunchHelper is implemented in both /Library/Java/JavaVirtualMachines/jdk1.8.0_77.jdk/Contents/Home/jre/bin/java (0x105abb4c0) and /Library/Java/JavaVirtualMachines/jdk1.8.0_77.jdk/Contents/Home/jre/lib/libinstrument.dylib (0x105b354e0). One of the two will be used. Which one is undefined.
# Warmup Iteration   1: 1208527.970 ops/s
# Warmup Iteration   2: 1436059.592 ops/s
Iteration   1: 1155912.403 ops/s
Iteration   2: 1158149.131 ops/s
Iteration   3: 1103212.762 ops/s

```

This branch:
```
# JMH version: 1.25
# VM version: JDK 1.8.0_77, Java HotSpot(TM) 64-Bit Server VM, 25.77-b03
# VM invoker: /Library/Java/JavaVirtualMachines/jdk1.8.0_77.jdk/Contents/Home/jre/bin/java
# VM options: -javaagent:/Users/dfa/Library/Application Support/JetBrains/Toolbox/apps/IDEA-U/ch-0/203.5981.155/IntelliJ IDEA.app/Contents/lib/idea_rt.jar=62647:/Users/dfa/Library/Application Support/JetBrains/Toolbox/apps/IDEA-U/ch-0/203.5981.155/IntelliJ IDEA.app/Contents/bin -Dfile.encoding=UTF-8
# Warmup: 2 iterations, 5 s each, 3 calls per op
# Measurement: 3 iterations, 10 s each, 4 calls per op
# Timeout: 10 min per iteration
# Threads: 1 thread, will synchronize iterations
# Benchmark mode: Throughput, ops/time
# Benchmark: benchmark.PropertyFetcherBenchMark.benchMarkThroughputDirectClassHierarchy

# Run progress: 0.00% complete, ETA 00:06:40
# Fork: 1 of 5
objc[13043]: Class JavaLaunchHelper is implemented in both /Library/Java/JavaVirtualMachines/jdk1.8.0_77.jdk/Contents/Home/jre/bin/java (0x10db2d4c0) and /Library/Java/JavaVirtualMachines/jdk1.8.0_77.jdk/Contents/Home/jre/lib/libinstrument.dylib (0x10dba74e0). One of the two will be used. Which one is undefined.
# Warmup Iteration   1: 4103756.324 ops/s
# Warmup Iteration   2: 6146158.084 ops/s
Iteration   1: 4577539.556 ops/s
Iteration   2: 4711435.476 ops/s
Iteration   3: 4654780.485 ops/s
```